requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
django-bootstrap4==2.3.1
django-cachalot
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-import-export==2.5.0
django-less==0.7.2
django-pivot==1.8.1
django-rq==2.5.1
django-static-precompiler==2.0
django-tables2==2.3.3
django-test-without-migrations==0.6
django-widget-tweaks==1.4.8
django==3.2
django-redis
factory-boy==3.2.0
gunicorn==20.1.0
idna==2.10
Markdown==3.2.2
newrelic==6.2.0.156
pandas==1.3.4
Pillow==8.0.1
protobuf==3.13.0
pylibmc
pymysql
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.1
requests-toolbelt==0.9.1
requests==2.26.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.26.6
whitenoise==5.2.0